pyproject.toml
[tool.poetry]
name = "payments_redsys"
version = "0.6.0"
description = "A django-payments backend for the Redsys payment gateway"
authors = ["Carles Barrobés <carles@barrobes.com>"]
package-mode = true
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"

django = ">=4.2.0"
django-payments = ">=3.0.1"
pydes = "^2.0.1"

[tool.poetry.group.dev.dependencies]
black = "~=24.0"
pyhamcrest = "^2.1.0"
pytest = "*"
pytest-cov = "*"
pytest-django = "~=4.8"
ruff = "^0.11.7"
wheel = "*"

[tool.pytest]
addopts = [
    "--capture=no",
    "--disable-warnings",
    "--no-migrations",
]

[tool.coverage.run]
omit = [
    # don't care about coverage of test code or migrations
    "**/test_*.py",
    "**/migrations/*",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
